Fix "Cannot convert an array" error
