+++ src/games/scrambled/Board/index.test.js
@@ -17,4 +17,17 @@ describe("scrambled Board component", () => {
     const tree = component.toJSON();
     expect(tree).toMatchSnapshot();
   });
+
+  it("renders waiting board if game not yet started", () => {
+    expect.hasAssertions();
+    const component = renderer.create(
+      <MemoryRouter>
+        <BoardGameContextMock overrideCtx={{ phase: "wait" }}>
+          <Board />
+        </BoardGameContextMock>
+      </MemoryRouter>
+    );
+    const tree = component.toJSON();
+    expect(tree).toMatchSnapshot();
+  });
 });
